requirements.txt
python-dotenv
pandas
numpy
librosa
filetype